# oa-console/openapi/oa-app.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18400",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/iam/grants": {
      "get": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "list_2",
        "parameters": [
          {
            "name": "subjectType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subjectId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListGrantRecord"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "grant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Grant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/elevations": {
      "post": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "elevate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Elevate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/elevation-requests/{id}/reject": {
      "post": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "rejectElevation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElevationDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/elevation-requests/{id}/approve": {
      "post": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "approveElevation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElevationDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/delegations": {
      "get": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "myDelegations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListDelegation"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "delegate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Delegate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/flow/todos/{taskId}/complete": {
      "post": {
        "tags": [
          "workbench-controller"
        ],
        "operationId": "complete",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/iam/elevation-requests": {
      "get": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "elevationRequests",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListRow"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/elevation-requests/mine": {
      "get": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "myElevationRequests",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListRow"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/flow/todos": {
      "get": {
        "tags": [
          "workbench-controller"
        ],
        "operationId": "myTodos",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListTodoView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flow/todos/mine": {
      "get": {
        "tags": [
          "workbench-controller"
        ],
        "operationId": "myApplications",
        "parameters": [
          {
            "name": "bizType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListMyApplicationView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/iam/grants/{grantId}": {
      "delete": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "revoke",
        "parameters": [
          {
            "name": "grantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/delegations/{id}": {
      "delete": {
        "tags": [
          "grant-controller"
        ],
        "operationId": "revokeDelegation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "Grant": {
        "required": [
          "roleId",
          "subjectId",
          "subjectType"
        ],
        "type": "object",
        "properties": {
          "subjectType": {
            "type": "string"
          },
          "subjectId": {
            "type": "string"
          },
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "scopeType": {
            "type": "string"
          },
          "scopeOrgIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "includeDescendants": {
            "type": "boolean"
          },
          "grantType": {
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time"
          },
          "validTo": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "Elevate": {
        "required": [
          "reason",
          "roleId"
        ],
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string"
          },
          "hours": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ElevationDecision": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "Delegate": {
        "required": [
          "delegateeUserId",
          "validTo"
        ],
        "type": "object",
        "properties": {
          "delegateeUserId": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "processKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "validFrom": {
            "type": "string",
            "format": "date-time"
          },
          "validTo": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "CompleteTask": {
        "required": [
          "outcome"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "onBehalfOf": {
            "type": "string"
          }
        }
      },
...
      "GrantRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tenantId": {
            "type": "integer",
            "format": "int64"
          },
          "subjectType": {
            "type": "string"
          },
          "subjectId": {
            "type": "string"
          },
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "scopeType": {
            "type": "string"
          },
          "scopeOrgIds": {
            "type": "string"
          },
          "includeDescendants": {
            "type": "boolean"
          },
          "grantType": {
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time"
          },
          "validTo": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "approvalInstanceId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "grantedBy": {
            "type": "string"
          },
          "grantedAt": {
            "type": "string",
            "format": "date-time"
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time"
          },
          "revokedBy": {
            "type": "string"
          },
          "revokeReason": {
            "type": "string"
          }
        }
      },
      "ResultListGrantRecord": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantRecord"
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "ResultListRow": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Row"
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "Row": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tenantId": {
            "type": "integer",
            "format": "int64"
          },
          "requesterId": {
            "type": "string"
          },
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "roleCode": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "hours": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "decidedBy": {
            "type": "string"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          },
          "decisionReason": {
            "type": "string"
          },
          "grantId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Delegation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tenantId": {
            "type": "integer",
            "format": "int64"
          },
          "delegatorUserId": {
            "type": "string"
          },
          "delegateeUserId": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "processKeys": {
            "type": "string"
          },
          "roleIds": {
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time"
          },
          "validTo": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultListDelegation": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Delegation"
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "ResultListTodoView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodoView"
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "TodoView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "taskId": {
            "type": "string"
          },
          "processInstanceId": {
            "type": "string"
          },
          "processDefinitionKey": {
            "type": "string"
          },
          "instanceId": {
            "type": "integer",
            "format": "int64"
          },
          "bizType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "applicantUserId": {
            "type": "string"
          },
          "applicantName": {
            "type": "string"
          },
          "assigneeUserId": {
            "type": "string"
          },
          "candidateGroup": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "orgId": {
            "type": "integer",
            "format": "int64"
          },
          "orgPath": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MyApplicationView": {
        "type": "object",
        "properties": {
          "bizType": {
            "type": "string"
          },
          "docNo": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "days": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultListMyApplicationView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyApplicationView"
            }
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}